+++ OpenRA.Mods.RA/NullLoadScreen.cs
@@ -24,7 +24,7 @@ namespace OpenRA.Mods.RA
 				return;
 
 			// Draw a black screen
-			Game.Renderer.BeginFrame(float2.Zero, 1f);
+			Game.Renderer.BeginFrame(int2.Zero, 1f);
 			Game.Renderer.EndFrame( new NullInputHandler() );
 		}
